senslarge/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file senslarge.egg-info/requirements.txt senslarge.egg-info/requirements.in
#
attrs==17.4.0 

            # via pytest
cairocffi==0.8.0          # via cairosvg, weasyprint
cairosvg==2.1.3           # via weasyprint
cffi==1.11.5              # via cairocffi, weasyprint
chardet==3.0.4            # via requests
coverage==4.5.1           # via pytest-cov
cssselect2==0.2.1         # via cairosvg, weasyprint
cssselect==1.0.3          # via pyquery
defusedxml==0.5.0         # via cairosvg
diff-cover==1.0.2
django-bootstrap-static==3.3.7.2
django-bootstrap3==9.1.0
django-braces==1.13.0
django-fontawesome==0.3.1
django-healthchecks==1.4.2
django-mailgun==0.9.1
django-ordered-model==1.4.3
django-redis==4.9.0
django-weasyprint==0.5.2



django==2.0.4             # via django-healthchecks, django-redis, django-weasyprint
djangorestframework==3.8.2
factory-boy==2.10.0
faker==0.8.13             # via factory-boy
flake8==3.5.0
html5lib==1.0.1           # via weasyprint
idna==2.6                 # via requests
inflect==0.2.5            # via jinja2-pluralize
inflection==0.3.1         # via pytest-factoryboy
jinja2-pluralize==0.3.0   # via diff-cover
jinja2==2.10              # via diff-cover, jinja2-pluralize
lxml==4.2.1               # via pyquery
markupsafe==1.0           # via jinja2
mccabe==0.6.1             # via flake8
more-itertools==4.1.0     # via pytest
pdfrw==0.4                # via weasyprint
pillow==5.1.0             # via cairosvg
pluggy==0.6.0             # via pytest
psycopg2==2.7.4
py==1.5.3                 # via pytest
pycodestyle==2.3.1        # via flake8
pycparser==2.18           # via cffi
pyflakes==1.6.0           # via flake8
pygal==2.4.0
pygments==2.2.0           # via diff-cover
pyphen==0.9.4             # via weasyprint
pyquery==1.4.0
pytest-cov==2.5.1
pytest-django==3.2.1
pytest-factoryboy==2.0.1
pytest==3.5.0
python-dateutil==2.7.2    # via faker
pytz==2018.4              # via django
pyyaml==3.12              # via django-fontawesome
raven==6.7.0
redis==2.10.6             # via django-redis
requests==2.18.4          # via django-healthchecks, django-mailgun
six==1.11.0               # via diff-cover, django-healthchecks, django-mailgun, faker, html5lib, more-itertools, pytest, python-dateutil
text-unidecode==1.2       # via faker
tinycss2==0.6.1           # via cairosvg, cssselect2, weasyprint
urllib3==1.22             # via requests
weasyprint==0.42.3        # via django-weasyprint
webencodings==0.5.1       # via html5lib, tinycss2
